unit tests from blocker.rs moved to tests/unit/blocker.rs
